requirements.txt
azure-functions
azure-identity
pandas
psycopg2-binary
